Skip already downloaded images

+++ fetch-from-nasa.py
@@ -1,4 +1,4 @@
-import json, re
+import json, re, os
 import requests
 from dotenv import dotenv_values
 from tqdm import tqdm
@@ -93,8 +93,9 @@ def get_page_items(items: list):
         else:
             metadata = get_metadata(item_id, file_links[-1]['href'], file_links[-1]['href'])
 
-        with open(f"web/{metadata['file']}", "wb") as file:
-            file.write(session.get(metadata['file_url'], headers = headers).content)
+        if not os.path.exists(f"web/{metadata['file']}"):
+            with open(f"web/{metadata['file']}", "wb") as file:
+                file.write(session.get(metadata['file_url'], headers = headers).content)
 
         collection['photos'].append(metadata)
